feat: 과목 리스트조회 & 검색 구현

+++ src/main/java/Dolphin/ShoppingCart/domain/course/entity/Teach.java
@@ -38,7 +38,7 @@ public class Teach extends BaseEntity {
     @Enumerated(EnumType.STRING)
     private SemesterType semester;
 
-    private Integer number;   // 과목번호
+    private Long number;   // 과목번호
     private String className; // 분반명
     private Integer targetGrade;
